+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - New #20: Add `sendDocument` method.
 - New #21: Add `sendVideo` method.
 - New #22: Add `getStarTransactions` method and `StarTransactions`, `StarTransaction`, `TransactionPartner`, 
-  `RevenueWithdrawalState` types. 
+  `RevenueWithdrawalState` types.
+- New #23: Add `logOut` method.
 
 ## 0.1.0 June 10, 2024